Moved NumberInput under Inputs folder
