+++ app/components/Inputs/NumberInput/NumberInput.scss
